Problem: interprocess communication
It's often a non-trivial problem to share data structures and send messages
between processes.

Solution: build a thin library to do just that

There are some very exciting projects like
[Flow-IPC](https://github.com/Flow-IPC/ipc) but they are currently not
suitable for our needs because of lack of arm64/macos support. Arguably,
this is fixable, but then its perceived build surface complexity comes
into the picture (it's a big library). It does have amazing benefits,
though, like transmitting C++ structures over the wire.

Oink is an attempt to make a simple library that works well enough, fast
enough. Most of workloads sent through this will be terminated by a
database operation (as the use case for this library is Omnigres), and
the hypothesis here is that the overhead of this IPC will be negligible
comparatively.

If this hypothesis won't be correct, we can try and make use of
Flow-IPC, after all. For now, the job is to make something workable that
Omnigres components can share.