Problem: message memory is not being freed
This is rather a significant problem as we'll
eventually run out of space.

Solution: count references
and deallocate when counter reaches zero.

The message will not be deallocated until it is
processed because the initial counter is set to 2
(to account for the sender and [pending] receiver)

This is a quick and dirty prototype.